renaming auth.js file for validation.js
